exif autorotate for proofs
Find below a patch which adds auto-rotation of proof images based on image
EXIF information.  It is a revision of a proof-of-concept patch I did up for
gqview 2.0.4 which, for well documented reasons, didn't get anywhere after
submission.  The patch below is against geeqie 1.0.

The proof auto-rotation feature is controlled by a new setting in the
"convenience" section of the image preferences dialog.  This way the
existing behaviour is still available for those who want to use it.

I find this new feature very useful when preparing proofing sheets and I
suspect others would too.  Please consider this for inclusion in geeqie.