Always scroll to source directory when walking up in list
When walking ones directory tree in list mode selecting "..", sometimes the views scrolls to the directory from where you came and sometimes it does not.
Attached is a small patch that makes sure the view always scrolls to the directory from where we came.
Apparently the problem was that the source directory entry name sometimes is "..", not very informative when searching the new directory contents after an entry with that name, so changed to search after path instead.

http://sourceforge.net/tracker/?func=detail&aid=3127409&group_id=222125&atid=1054682